# openapi.yaml
openapi: 3.0.3
info:
  title: Notes API
  version: 1.0.0
  description: Simple JWT-protected notes API (users, auth, notes)
servers:
  - url: http://localhost:{port}
    variables: 
      port: 
        default: "3000"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: 
          type: string
    User:
      type: object
      properties:
        id: 
          type: integer
        email: 
          type: string
          format: email
        created_at: 
          type: string
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: 
          type: string
          format: email
        password: 
          type: string
          minLength: 6
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: 
          type: string
          format: email
        password: 
          type: string
    AuthToken:
      type: object
      properties:
        token: 
          type: string
        message:
          type: string
        userId:
          type: integer
    Note:
      type: object
      properties:
        id: 
          type: integer
        user_id: 
          type: integer
        content: 
          type: string
        created_at: 
          type: string
    NoteCreate:
      type: object
      required: [content]
      properties:
        content: 
          type: string
          maxLength: 1000
    NotesList:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties: 
                  ok: 
                    type: boolean
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json: 
            schema: 
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json: 
              schema: 
                type: object
                properties:
                  message:
                    type: string
                  userId:
                    type: integer
        "400": 
          description: Bad Request
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json: 
            schema: 
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json: 
              schema: 
                $ref: "#/components/schemas/AuthToken"
        "401": 
          description: Unauthorized
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/Error"
  /notes:
    get:
      summary: List user's notes
      security: 
        - bearerAuth: []
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotesList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteCreate"
      responses:
        "201":
          description: Note created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  noteId:
                    type: integer
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /notes/{id}:
    delete:
      summary: Delete a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"